requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
peft>=0.10.0
trl>=0.9.4
accelerate>=0.30.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
protobuf>=4.25.3
evaluate>=0.4.1
tqdm
numpy
pandas
python-dotenv
pillow
python-docx
pdfminer.six
pandas
openpyxl
extract-msg
tqdm
transformers
torch
datasets
peft
trl
pytesseract
